+++ client/src/languageFeatures/internal/brunoFiles/completionItems/provideTsLangCompletionItems.ts
@@ -12,12 +12,7 @@ import {
     mapFromVsCodePosition,
     TypedCollectionItemProvider,
 } from "@shared";
-import {
-    RequestFileBlockName,
-    Block,
-    getFirstParameterForInbuiltFunctionIfStringLiteral,
-    getInbuiltFunctionIdentifiers,
-} from "@global_shared";
+import { RequestFileBlockName, Block } from "@global_shared";
 import { getPositionWithinTempJsFile } from "../shared/codeBlocksUtils/getPositionWithinTempJsFile";
 import { mapToRangeWithinBruFile } from "../shared/codeBlocksUtils/mapToRangeWithinBruFile";
 import { getRequestFileDocumentSelector } from "../shared/getRequestFileDocumentSelector";
@@ -26,7 +21,6 @@ import {
     TempJsSyncRequest,
     waitForTempJsFileToBeInSync,
 } from "../../shared/temporaryJsFilesUpdates/external/waitForTempJsFileToBeInSync";
-import { mapToEnvVarNameParams } from "../shared/codeBlocksUtils/mapToGetEnvVarNameParams";
 import { getCodeBlockContainingPosition } from "../shared/codeBlocksUtils/getCodeBlockContainingPosition";
 
 type CompletionItemRange =
@@ -67,27 +61,6 @@ export function provideTsLangCompletionItems(
                 if (token.isCancellationRequested) {
                     addLogEntryForCancellation(logger);
                     return undefined;
-                }
-
-                const envVariableResult =
-                    getFirstParameterForInbuiltFunctionIfStringLiteral(
-                        mapToEnvVarNameParams(
-                            {
-                                file: {
-                                    collection,
-                                    blockContainingPosition,
-                                },
-                                request: { document, position, token },
-                                logger,
-                            },
-                            getInbuiltFunctionIdentifiers(),
-                        ),
-                    );
-
-                if (envVariableResult) {
-                    // Completions for environment variables are already provided via the language server.
-                    // Since they are also provided in `.js` files, avoid fetching completions via temp js file in this case.
-                    return;
                 }
 
                 return getResultsViaTempJsFile(